refactor: SpecialTileTypeをsrc/types配下に移動
