+++ src/game/map.ts
@@ -15,10 +15,15 @@ import {
 	MAP_WIDTH,
 	STAIRS_COUNT,
 } from "../constants";
-import type { GameMap, Position, Tile, TileType } from "../types";
+import type {
+	GameMap,
+	Position,
+	SpecialTileType,
+	Tile,
+	TileType,
+} from "../types";
 import type { RNG } from "../utils/rng";
 import { generateBSPMap, type Room } from "./bsp";
-import type { SpecialTileType } from "./tileEffect";
 
 /**
  * 座標がマップ範囲内かを判定
